+++ internal/store/token.go
@@ -19,11 +19,11 @@ func NewPostgresTokenStore(db *sql.DB) *PostgresTokenStore {
 
 type TokenStore interface {
 	Insert(token *tokens.Token) error
-	CreateNewToken(userID int64, ttl time.Duration, scope string) (*tokens.Token, error)
-	DeleteAllTokensForUser(userID int64, scope string) error
+	CreateNewToken(userID int, ttl time.Duration, scope string) (*tokens.Token, error)
+	DeleteAllTokensForUser(userID int, scope string) error
 }
 
-func (t *PostgresTokenStore) CreateNewToken(userID int64, ttl time.Duration, scope string) (*tokens.Token, error) {
+func (t *PostgresTokenStore) CreateNewToken(userID int, ttl time.Duration, scope string) (*tokens.Token, error) {
 	token, err := tokens.GenerateToken(userID, ttl, scope)
 	if err != nil {
 		return nil, err
@@ -43,7 +43,7 @@ func (t *PostgresTokenStore) Insert(token *tokens.Token) error {
 	return err
 }
 
-func (t *PostgresTokenStore) DeleteAllTokensForUser(userID int64, scope string) error {
+func (t *PostgresTokenStore) DeleteAllTokensForUser(userID int, scope string) error {
 	query := `
 	DELETE FROM tokens
 	WHERE user_id = $1 and scope = $2
